+++ include/serial.h
@@ -6,6 +6,6 @@
 int init_serial();
 void write_serial(char a);
 void serial_print(const char *str);
-void serial_print_hex(uint32_t n);
+void serial_print_hex(uint64_t n);
 
 #endif
